+++ contracts/mocks/token/ERC20/facets/ERC20BurnableFacetMock.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.17;
+pragma solidity 0.8.19;
 
 import {IForwarderRegistry} from "./../../../../metatx/interfaces/IForwarderRegistry.sol";
 import {ERC20BurnableFacet} from "./../../../../token/ERC20/facets/ERC20BurnableFacet.sol";