+++ contracts/mocks/token/ERC721/ERC721ReceiverMock.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.17;
+pragma solidity 0.8.19;
 
 import {IERC721Receiver} from "./../../../token/ERC721/interfaces/IERC721Receiver.sol";
 import {ERC721Storage} from "./../../../token/ERC721/libraries/ERC721Storage.sol";